requirements.txt
flask
python-dotenv
transformers
torch
PyPDF2
python-docx
gTTS
werkzeug